telegram_parser/requirements.txt
aio-pika
tgcrypto
pyrofork
python-dotenv